Free-Game-Notifier/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "trafilatura>=2.0.0",
]
